* #988164: Matrix: Auto-scrolling so fast you can't see sweep distance   Tweak the values so as to start up more slowly and smoothly
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@6075 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ gui/rosegardencanvasview.h
@@ -119,7 +119,7 @@ protected:
 
     bool m_smoothScroll;
     int m_smoothScrollTimeInterval;
-    int m_minDeltaScroll;
+    float m_minDeltaScroll;
     QTime m_scrollTimer;
     QTime m_scrollAccelerationTimer;
     static const int DefaultSmoothScrollTimeInterval;
